Added spheres and hittable objects to the code

+++ RayTracer/src/library/vec3.h
@@ -76,19 +76,19 @@ inline vec3 operator*(const vec3& u, const vec3& v) {
 	return vec3(u.e[0] * v.e[0], u.e[1] * v.e[1], u.e[2] * v.e[2]);
 }
 
-inline vec3 operator*(double t, const vec3& v) {
+inline vec3 operator*(float t, const vec3& v) {
 	return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
-inline vec3 operator*(const vec3& v, double t) {
+inline vec3 operator*(const vec3& v, float t) {
 	return t * v;
 }
 
-inline vec3 operator/(vec3 v, double t) {
+inline vec3 operator/(vec3 v, float t) {
 	return (1 / t) * v;
 }
 
-inline double dot(const vec3& u, const vec3& v) {
+inline float dot(const vec3& u, const vec3& v) {
 	return u.e[0] * v.e[0]
 		+ u.e[1] * v.e[1]
 		+ u.e[2] * v.e[2];
